render settings inline for inline components
